refactor & fix & style
