+++ src/components/common/Header/BurgerLoggedInHeader.tsx
@@ -9,18 +9,18 @@ export default function BurgerLoggedInHeader({ isOpen }: { isOpen: boolean }) {
       <nav className='navigation-auth-burger'>
         <ul className='navigation-auth-burger__list list-reset'>
           <li className='navigation-auth-burger__list-element'>
-            <NavLink className='navigation-auth-burger__link link-hover link-active' to='/'>Главная</NavLink>
+            <NavLink className='navigation-auth-burger__link link-hover active-underline' to='/'>Главная</NavLink>
           </li>
           <li className='navigation-auth-burger__list-element'>
-            <NavLink className='navigation-auth-burger__link link-hover link-active' to='/movies'>Фильмы</NavLink>
+            <NavLink className='navigation-auth-burger__link link-hover active-underline' to='/movies'>Фильмы</NavLink>
           </li>
           <li className='navigation-auth-burger__list-element'>
-            <NavLink className='navigation-auth-burger__link link-hover link-active' to='/saved-movies'>Сохраненные фильмы</NavLink>
+            <NavLink className='navigation-auth-burger__link link-hover active-underline' to='/saved-movies'>Сохраненные фильмы</NavLink>
           </li>
         </ul>
         <NavLink className='navigation-auth-burger__link link-hover' to='/profile'>
           <button type='button' className='navigation-auth-burger__account-btn btn-reset'>
-            <span className='navigation-auth-burger__account-btn-text link-active'>Аккаунт</span>
+            <span className='navigation-auth-burger__account-btn-text active-underline'>Аккаунт</span>
             <span className='navigation-auth-burger__account-btn-icon' />
           </button>
         </NavLink>
