+++ src/components/others/BurgerBtn/BurgerBtn.css
@@ -20,6 +20,8 @@
   flex-direction: column;
   justify-content: space-between;
   align-items: center;
+
+  cursor: pointer;
 }
 
 .burger-menu__line {
