+++ src/components/others/Logo/Logo.tsx
@@ -1,0 +1,14 @@
+import './Logo.css';
+
+import React from 'react';
+import { Link } from 'react-router-dom';
+
+import logo from '../../../images/logo.svg';
+
+export default function Logo() {
+  return (
+    <Link to='/' className='logo'>
+      <img src={logo} alt='Логотип' className='logo__img btn-hover active-btn-effect' />
+    </Link>
+  );
+}
