+++ src/components/others/PortfolioLinkElement/PortfolioLinkElement.css
@@ -1,25 +1,3 @@
-.portfolio {
-  padding: 50px 70px 125px;
-  background-color: var(--background-color-light-black);
-}
-
-.portfolio__title {
-  color: #8b8b8b;
-  font-feature-settings: "clig" off, "liga" off;
-  font-size: 18px;
-  font-style: normal;
-  font-weight: 400;
-  line-height: 20px; /* 111.111% */
-
-  margin: 0 0 50px;
-}
-
-.portfolio__list {
-  display: flex;
-  flex-direction: column;
-  gap: 20px;
-}
-
 .portfolio__list-element {
   display: flex;
   justify-content: space-between;
@@ -53,21 +31,12 @@
 
   text-decoration: none;
 
-  /* подлежит правке */
   display: flex;
   justify-content: space-between;
   width: 100%;
 }
 
 @media screen and (max-width: 768px) {
-  .portfolio {
-    padding: 45px 50px 90px;
-  }
-
-  .portfolio__title {
-    color: #a0a0a0;
-  }
-
   .portfolio__list-element-text {
     font-size: 28px;
     line-height: 50px;
@@ -82,15 +51,6 @@
 }
 
 @media screen and (max-width: 400px) {
-  .portfolio {
-    padding: 35px 14px 70px;
-  }
-
-  .portfolio__title {
-    font-size: 14px;
-    margin-bottom: 40px;
-  }
-
   .portfolio__list-element-text {
     font-size: 18px;
     line-height: 28px;
