+++ src/components/others/SectionTitleRow/SectionTitleRow.css
@@ -1,0 +1,27 @@
+.section__title {
+  color: var(--base-color-of-text);
+  font-size: 22px;
+  font-style: normal;
+  font-weight: 400;
+  line-height: normal;
+  letter-spacing: -0.88px;
+
+  margin: 0 0 23px;
+}
+
+.section__row {
+  width: 100%;
+  height: 1px;
+  background-color: #dadada;
+
+  margin: 0 0 66px;
+}
+
+@media screen and (max-width: 400px) {
+  .about-me__title {
+    font-size: 18px;
+    letter-spacing: -0.72px;
+
+    margin-bottom: 28px;
+  }
+}
