+++ src/components/pages/Main/Main.tsx
@@ -1,11 +1,11 @@
 import './Main.css';
 
 import React from 'react';
-import Promo from '../../others/Promo/Promo';
-import AboutProject from '../../others/AboutProject/AboutProject';
-import Techs from '../../others/Techs/Techs';
-import AboutMe from '../../others/AboutMe/AboutMe';
-import Portfolio from '../../others/Portfolio/Portfolio';
+import Promo from '../../sections/Promo/Promo';
+import AboutProject from '../../sections/AboutProject/AboutProject';
+import Techs from '../../sections/Techs/Techs';
+import AboutMe from '../../sections/AboutMe/AboutMe';
+import Portfolio from '../../sections/Portfolio/Portfolio';
 
 export default function Main() {
   return (
