+++ src/components/pages/Profile/Profile.css
@@ -66,19 +66,9 @@
 
   width: 100%;
   max-width: 350px;
-
-  transition: color 0.2s linear;
-}
-
-.page-profile__input:hover {
-  color: #ee3465;
 }
 
 .page-profile__input:disabled:hover {
-  color: var(--base-color-of-text);
-}
-
-.page-profile__input:focus {
   color: var(--base-color-of-text);
 }
 
@@ -100,10 +90,12 @@
 .page-profile__btns {
   display: flex;
   flex-direction: column;
+  align-items: center;
   gap: 16px;
 }
 
 .page-profile__btn-edit {
+  width: fit-content;
   padding: 0 5px 0;
 
   color: var(--base-color-of-text);
@@ -115,6 +107,7 @@
 }
 
 .page-profile__btn-logout {
+  width: fit-content;
   padding: 0 5px 0;
 
   color: #ee3465;
