+++ src/components/sections/Techs/Techs.css
@@ -3,25 +3,6 @@
   background-color: var(--background-color-grey);
 }
 
-.techs__title {
-  color: var(--base-color-of-text);
-  font-size: 22px;
-  font-style: normal;
-  font-weight: 400;
-  line-height: normal;
-  letter-spacing: -0.88px;
-
-  margin: 0 0 23px;
-}
-
-.techs__row {
-  width: 100%;
-  height: 1px;
-  background-color: #dadada;
-
-  margin-bottom: 90px;
-}
-
 .techs__subtitle {
   color: var(--base-color-of-text);
   text-align: center;
@@ -31,7 +12,7 @@
   line-height: 58px; /* 116% */
   letter-spacing: -2px;
 
-  margin: 0 0 26px;
+  margin: 90px 0 26px;
 }
 
 .techs__text {
@@ -89,11 +70,8 @@
     padding: 90px 50px;
   }
 
-  .techs__row {
-    margin-bottom: 80px;
-  }
-
   .techs__subtitle {
+    margin-top: 80px;
     margin-bottom: 22px;
   }
 
@@ -115,22 +93,12 @@
     padding: 70px 18px;
   }
 
-  .techs__title {
-    font-size: 18px;
-    letter-spacing: -0.72px;
-
-    margin-bottom: 28px;
-  }
-
-  .techs__row {
-    margin-bottom: 60px;
-  }
-
   .techs__subtitle {
     font-size: 30px;
     line-height: normal;
     letter-spacing: -1.2px;
 
+    margin-top: 60px;
     margin-bottom: 24px;
   }
